# HDDS-4933. Make CI checks fail faster