… hadoop-ozone/dist/src/main/compose/ozone-mr/test.sh
… hadoop-ozone/dist/src/main/compose/test-all.sh
… hadoop-ozone/dist/src/main/compose/testlib.sh
… hadoop-ozone/dist/src/main/compose/upgrade/test.sh
… hadoop-ozone/dist/src/test/shell/compose_testlib.bats