## What changes were proposed in this pull request?

Currently most CI checks run to completion even if other checks fail.  This is useful if we want to find as many problems as possible in a single CI run.  However, it is a waste of resources when trying to get a clean run before merge.

I propose to make checks fail faster for PR runs, but keep existing behavior for `push` and `schedule` builds.

https://issues.apache.org/jira/browse/HDDS-4933

## How was this patch tested?

1. Introduced unit test failure, verified that test execution is stopped after failed test if `FAIL_FAST=true`, but continues with unrelated modules if `false`:

```
$ FAIL_FAST=true ./hadoop-ozone/dev-support/checks/unit.sh
...
[INFO] Reactor Summary for Apache Hadoop Ozone Main 1.1.0-SNAPSHOT:
[INFO]
[INFO] Apache Hadoop Ozone Main ........................... SUCCESS [  0.824 s]
[INFO] Apache Hadoop HDDS ................................. SUCCESS [  0.360 s]
[INFO] Apache Hadoop HDDS Hadoop Client dependencies ...... SUCCESS [  0.816 s]
[INFO] Apache Hadoop HDDS Hadoop Test dependencies ........ SUCCESS [  0.099 s]
[INFO] Apache Hadoop HDDS Hadoop Server dependencies ...... SUCCESS [  0.892 s]
[INFO] Apache Hadoop HDDS Client Interface ................ SUCCESS [  3.945 s]
[INFO] Apache Hadoop HDDS Admin Interface ................. SUCCESS [  1.165 s]
[INFO] Apache Hadoop HDDS Server Interface ................ SUCCESS [  1.361 s]
[INFO] Apache Hadoop HDDS Test Utils ...................... SUCCESS [  0.165 s]
[INFO] Apache Hadoop HDDS Config .......................... FAILURE [  2.556 s]
[INFO] Apache Hadoop HDDS Common .......................... SKIPPED
[INFO] Apache Hadoop HDDS Client .......................... SKIPPED
[INFO] Apache Hadoop HDDS Server Framework ................ SKIPPED
[INFO] Apache Ozone/HDDS Documentation .................... SKIPPED
[INFO] Apache Hadoop HDDS Container Service ............... SKIPPED
[INFO] Apache Hadoop HDDS SCM Server ...................... SKIPPED
[INFO] Apache Hadoop HDDS Tools ........................... SKIPPED
[INFO] Apache Hadoop Ozone ................................ SKIPPED
[INFO] Apache Hadoop Ozone Client Interface ............... SKIPPED
[INFO] Apache Hadoop Ozone Common ......................... SKIPPED
[INFO] Apache Hadoop Ozone Storage Interface .............. SKIPPED
[INFO] Apache Hadoop Ozone Client ......................... SKIPPED
[INFO] Apache Hadoop Ozone Manager Server ................. SKIPPED
[INFO] Apache Hadoop Ozone FileSystem Common .............. SKIPPED
[INFO] Apache Hadoop Ozone FileSystem ..................... SKIPPED
[INFO] Apache Hadoop Ozone Recon CodeGen .................. SKIPPED
[INFO] Apache Hadoop Ozone Recon .......................... SKIPPED
[INFO] Apache Hadoop Ozone Tools .......................... SKIPPED
[INFO] Apache Hadoop Ozone S3 Gateway ..................... SKIPPED
[INFO] Apache Hadoop Ozone CSI service .................... SKIPPED
[INFO] Apache Hadoop Ozone Datanode ....................... SKIPPED
[INFO] Apache Hadoop Ozone Insight Tool ................... SKIPPED
[INFO] Apache Hadoop Ozone FileSystem Shaded .............. SKIPPED
[INFO] Apache Hadoop Ozone FS Hadoop 2.x compatibility .... SKIPPED
[INFO] Apache Hadoop Ozone FS Hadoop 3.x compatibility .... SKIPPED
[INFO] Apache Hadoop Ozone Distribution ................... SKIPPED
[INFO] Apache Hadoop Ozone Fault Injection Tests .......... SKIPPED
[INFO] Apache Hadoop Ozone Network Tests .................. SKIPPED
[INFO] ------------------------------------------------------------------------
[INFO] BUILD FAILURE
```

```
$ ./hadoop-ozone/dev-support/checks/unit.sh
...
[INFO] Reactor Summary for Apache Hadoop Ozone Main 1.1.0-SNAPSHOT:
[INFO]
[INFO] Apache Hadoop Ozone Main ........................... SUCCESS [  0.664 s]
[INFO] Apache Hadoop HDDS ................................. SUCCESS [  0.364 s]
[INFO] Apache Hadoop HDDS Hadoop Client dependencies ...... SUCCESS [  0.766 s]
[INFO] Apache Hadoop HDDS Hadoop Test dependencies ........ SUCCESS [  0.115 s]
[INFO] Apache Hadoop HDDS Hadoop Server dependencies ...... SUCCESS [  0.795 s]
[INFO] Apache Hadoop HDDS Client Interface ................ SUCCESS [  3.827 s]
[INFO] Apache Hadoop HDDS Admin Interface ................. SUCCESS [  1.168 s]
[INFO] Apache Hadoop HDDS Server Interface ................ SUCCESS [  1.569 s]
[INFO] Apache Hadoop HDDS Test Utils ...................... SUCCESS [  0.158 s]
[INFO] Apache Hadoop HDDS Config .......................... FAILURE [  2.487 s]
[INFO] Apache Hadoop HDDS Common .......................... SKIPPED
[INFO] Apache Hadoop HDDS Client .......................... SKIPPED
[INFO] Apache Hadoop HDDS Server Framework ................ SKIPPED
[INFO] Apache Ozone/HDDS Documentation .................... SUCCESS [  3.413 s]
[INFO] Apache Hadoop HDDS Container Service ............... SKIPPED
[INFO] Apache Hadoop HDDS SCM Server ...................... SKIPPED
[INFO] Apache Hadoop HDDS Tools ........................... SKIPPED
[INFO] Apache Hadoop Ozone ................................ SUCCESS [  0.013 s]
[INFO] Apache Hadoop Ozone Client Interface ............... SUCCESS [  3.422 s]
[INFO] Apache Hadoop Ozone Common ......................... SKIPPED
[INFO] Apache Hadoop Ozone Storage Interface .............. SKIPPED
[INFO] Apache Hadoop Ozone Client ......................... SKIPPED
[INFO] Apache Hadoop Ozone Manager Server ................. SKIPPED
[INFO] Apache Hadoop Ozone FileSystem Common .............. SKIPPED
[INFO] Apache Hadoop Ozone FileSystem ..................... SKIPPED
[INFO] Apache Hadoop Ozone Recon CodeGen .................. SKIPPED
[INFO] Apache Hadoop Ozone Recon .......................... SKIPPED
[INFO] Apache Hadoop Ozone Tools .......................... SKIPPED
[INFO] Apache Hadoop Ozone S3 Gateway ..................... SKIPPED
[INFO] Apache Hadoop Ozone CSI service .................... SKIPPED
[INFO] Apache Hadoop Ozone Datanode ....................... SKIPPED
[INFO] Apache Hadoop Ozone Insight Tool ................... SKIPPED
[INFO] Apache Hadoop Ozone FileSystem Shaded .............. SKIPPED
[INFO] Apache Hadoop Ozone FS Hadoop 2.x compatibility .... SKIPPED
[INFO] Apache Hadoop Ozone FS Hadoop 3.x compatibility .... SKIPPED
[INFO] Apache Hadoop Ozone Distribution ................... SKIPPED
[INFO] Apache Hadoop Ozone Fault Injection Tests .......... SUCCESS [  0.014 s]
[INFO] Apache Hadoop Ozone Network Tests .................. SUCCESS [  0.227 s]
[INFO] ------------------------------------------------------------------------
[INFO] BUILD FAILURE
```

2. Verified similar behavior with failing acceptance tests:

```
$ OZONE_ACCEPTANCE_SUITE=failing ./hadoop-ozone/dev-support/checks/acceptance.sh
...
ERROR: Test execution of failing1 is FAILED!!!!
...
ERROR: Test execution of failing2 is FAILED!!!!
```

```
$ FAIL_FAST=true OZONE_ACCEPTANCE_SUITE=failing ./hadoop-ozone/dev-support/checks/acceptance.sh
...
ERROR: Test execution of failing1 is FAILED!!!!
```

Regular CI on `push` in fork:
https://github.com/adoroszlai/hadoop-ozone/actions/runs/646130999